requirements.txt
RPi.GPIO
pyserial
opencv-python
numpy
onnxruntime